Pull out common pixel render iteration utility
